Make sure nginx is installed before proceeding with install.

+++ onionimbus_node/INSTALL.sh
@@ -4,18 +4,27 @@ if [[ $EUID -ne 0 ]]; then
   echo "This script must be run as root" 1>&2
   exit 1
 fi
-
-# Create the directory for all of our configurations
-mkdir /etc/onionimbus
-
-# Run this script every minute
-crontab -l > mycron
-echo "* * * * * /etc/onionimbus/everyminute.sh" >> mycron
-crontab mycron
-rm mycron
-
-# Deply configuration files
-mv ./conf/nginx.conf /etc/onionimbus/nginx.conf
-ln -s /etc/onionimbus/nginx.conf /etc/nginx/sites-enabled/onionimbus
-# Move shell scripts
-mv ./outside_webroot/everyminute.sh /etc/onionimbus/everyminute.sh
+if [ -d /etc/nginx ] 
+then
+  # nginx is installed -- hooray!
+  
+  # Create the directory for all of our configurations
+  mkdir /etc/onionimbus
+  
+  # Run this script every minute
+  crontab -l > mycron
+  echo "* * * * * /etc/onionimbus/everyminute.sh" >> mycron
+  crontab mycron
+  rm mycron
+  
+  # Deply configuration files
+  mv ./conf/nginx.conf /etc/onionimbus/nginx.conf
+  ln -s /etc/onionimbus/nginx.conf /etc/nginx/sites-enabled/onionimbus
+  # Move shell scripts
+  mv ./outside_webroot/everyminute.sh /etc/onionimbus/everyminute.sh
+else 
+  # nginx not installed
+  echo "Please install nginx before you continue!" 1>&2
+  echo "sudo apt-get install update && sudo apt-get install nginx" 1 > &2
+  exit 1  
+fi
